Improve Sphinx Infrastructure
- Mandate Python >=3.10 across the repo (Not really enforced)
  - Older versions are not tested and Python isn't required for OpenDDS
    users. This is also the version in the last Ubuntu LTS release.
- Require Sphinx >=7.2
  - Mostly for some minor API improvements, but since `build.py` will
    automatically download it, there's not much of a reason to hold it
    back.
- Try removing workaround for a RTD error.
- Workaround "LaTeX Error: Too deeply nested." on local PDF builds.
- Sphinx CMake domain:
  - Refactor to use a reusable custom domain framework
  - Add examples for all the directives and roles.
  - Rename `:nocontententry:` and other options for Sphinx 7.
